Task 3: Edit a Weather part 2
After clicking the [Edit Weather] button in the form, send a PUT request to the server to modify the location, temperature and the date of the changed item. After the successful request, fetch the items again and see that the changes have been made. After that, the [Edit Weather] button is deactivated and the [Add Weather] one is activated.