Task 3.3.2: Edit a Vacation
After clicking the [Edit Vacation] button in the form, send a PUT request to the server to modify the name, number of days and the date of the changed item. After the successful request, fetch the items again and see that the changes have been made. After that, the [Edit Vacation] button is deactivated and the [Add Vacation] one is activated.